rename control_plane subpackage to remote
