+++ flytekit/remote/nodes.py
@@ -13,9 +13,6 @@ from flytekit.common.exceptions import user as _user_exceptions
 from flytekit.common.mixins import artifact as _artifact_mixin
 from flytekit.common.mixins import hash as _hash_mixin
 from flytekit.common.utils import _dnsify
-from flytekit.control_plane import component_nodes as _component_nodes
-from flytekit.control_plane import identifier as _identifier
-from flytekit.control_plane.tasks.executions import FlyteTaskExecution
 from flytekit.core.context_manager import FlyteContextManager
 from flytekit.core.promise import NodeOutput
 from flytekit.core.type_engine import TypeEngine
@@ -26,6 +23,9 @@ from flytekit.models import node_execution as _node_execution_models
 from flytekit.models import task as _task_model
 from flytekit.models.core import execution as _execution_models
 from flytekit.models.core import workflow as _workflow_model
+from flytekit.remote import component_nodes as _component_nodes
+from flytekit.remote import identifier as _identifier
+from flytekit.remote.tasks.executions import FlyteTaskExecution
 
 
 class FlyteNode(_hash_mixin.HashOnReferenceMixin, _workflow_model.Node):
@@ -35,8 +35,8 @@ class FlyteNode(_hash_mixin.HashOnReferenceMixin, _workflow_model.Node):
         upstream_nodes,
         bindings,
         metadata,
-        flyte_task: "flytekit.control_plane.tasks.task.FlyteTask" = None,
-        flyte_workflow: "flytekit.control_plane.workflow.FlyteWorkflow" = None,
+        flyte_task: "flytekit.remote.tasks.task.FlyteTask" = None,
+        flyte_workflow: "flytekit.remote.workflow.FlyteWorkflow" = None,
         flyte_launch_plan=None,
         flyte_branch=None,
         parameter_mapping=True,
@@ -167,11 +167,11 @@ class FlyteNodeExecution(_node_execution_models.NodeExecution, _artifact_mixin.E
         self._interface = None
 
     @property
-    def task_executions(self) -> List["flytekit.control_plane.tasks.executions.FlyteTaskExecution"]:
+    def task_executions(self) -> List["flytekit.remote.tasks.executions.FlyteTaskExecution"]:
         return self._task_executions or []
 
     @property
-    def subworkflow_node_executions(self) -> Dict[str, "flytekit.control_plane.nodes.FlyteNodeExecution"]:
+    def subworkflow_node_executions(self) -> Dict[str, "flytekit.remote.nodes.FlyteNodeExecution"]:
         return (
             {}
             if self._subworkflow_node_executions is None
@@ -277,14 +277,14 @@ class FlyteNodeExecution(_node_execution_models.NodeExecution, _artifact_mixin.E
         )
 
     @property
-    def interface(self) -> "flytekit.control_plane.interface.TypedInterface":
+    def interface(self) -> "flytekit.remote.interface.TypedInterface":
         """
         Return the interface of the task or subworkflow associated with this node execution.
         """
         if self._interface is None:
 
-            from flytekit.control_plane.tasks.task import FlyteTask
-            from flytekit.control_plane.workflow import FlyteWorkflow
+            from flytekit.remote.tasks.task import FlyteTask
+            from flytekit.remote.workflow import FlyteWorkflow
 
             if not self.metadata.is_parent_node:
                 # if not a parent node, assume a task execution node
