+++ flytekit/remote/tasks/task.py
@@ -1,13 +1,13 @@
 from flytekit.common.exceptions import scopes as _exception_scopes
 from flytekit.common.exceptions import user as _user_exceptions
 from flytekit.common.mixins import hash as _hash_mixin
-from flytekit.control_plane import identifier as _identifier
-from flytekit.control_plane import interface as _interfaces
 from flytekit.engines.flyte import engine as _flyte_engine
 from flytekit.models import common as _common_model
 from flytekit.models import task as _task_model
 from flytekit.models.admin import common as _admin_common
 from flytekit.models.core import identifier as _identifier_model
+from flytekit.remote import identifier as _identifier
+from flytekit.remote import interface as _interfaces
 
 
 class FlyteTask(_hash_mixin.HashOnReferenceMixin, _task_model.TaskTemplate):
