+++ flytekit/remote/workflow.py
@@ -5,13 +5,13 @@ from flytekit.common.exceptions import scopes as _exception_scopes
 from flytekit.common.exceptions import system as _system_exceptions
 from flytekit.common.exceptions import user as _user_exceptions
 from flytekit.common.mixins import hash as _hash_mixin
-from flytekit.control_plane import identifier as _identifier
-from flytekit.control_plane import interface as _interfaces
-from flytekit.control_plane import nodes as _nodes
 from flytekit.engines.flyte import engine as _flyte_engine
 from flytekit.models import task as _task_models
 from flytekit.models.core import identifier as _identifier_model
 from flytekit.models.core import workflow as _workflow_models
+from flytekit.remote import identifier as _identifier
+from flytekit.remote import interface as _interfaces
+from flytekit.remote import nodes as _nodes
 
 
 class FlyteWorkflow(_hash_mixin.HashOnReferenceMixin, _workflow_models.WorkflowTemplate):
